refactors library into one source file per primitive radius function
