+++ src/access_reject.js
@@ -1,0 +1,18 @@
+
+module.exports = function access_reject(
+  decoded,
+  rinfo,
+  attributes,
+  vendor_attributes,
+  on_rejected
+) {
+  this.respond(
+    'auth',
+    decoded,
+    'Access-Reject',
+    rinfo,
+    attributes,
+    vendor_attributes,
+    on_rejected || function() {}
+  )
+}
