+++ src/accounting_respond.js
@@ -1,0 +1,18 @@
+
+module.exports = function accounting_respond(
+  decoded,
+  rinfo,
+  attributes,
+  vendor_attributes,
+  on_responded
+) {
+  this.respond(
+    'acct',
+    decoded,
+    'Accounting-Response',
+    rinfo,
+    attributes,
+    vendor_attributes,
+    on_responded
+  )
+}
