+++ src/send.js
@@ -1,0 +1,11 @@
+
+module.exports = function send(buffer, rinfo, on_sent) {
+  this.send(
+    buffer,
+    0,
+    buffer.length,
+    rinfo.port,
+    rinfo.address,
+    on_sent
+  )
+}
